# mycute-go/src/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "## API概要\\nMYCUTE REST APIを定義する。\\nURL最大長のリスクを避ける為、検索は query parameter ではなく body json を使用する。\\nGin に SEARCH/QUERY method の実装がないため、検索を POST にて行う。",
        "title": "MYCUTE",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
...
        "/v1/cubes/absorb": {
            "put": {
                "description": "- USR によってのみ使用できる\n- Cube に知識を追加する\n- 実行には AbsorbLimit に残数が必要",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "v1 Cube"
                ],
                "summary": "コンテンツを取り込む",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Bearer ??????????",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AbsorbCubeParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/AbsorbCubeRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            }
        },
        "/v1/cubes/create": {
            "post": {
                "description": "- USR によってのみ使用できる\n- 作成者は「神権限 (Limit = 0: 無制限)」を持つ\n- Cube は知識ベースとして機能し、Absorb/Memify/Search を通じて利用される",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "v1 Cube"
                ],
                "summary": "新しいCubeを作成する。",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Bearer ??????????",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateCubeParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/CreateCubeRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            }
        },
        "/v1/cubes/delete": {
            "delete": {
                "description": "- USR によってのみ使用できる\n- 指定したCubeと関連データを完全に削除する",
                "tags": [
                    "v1 Cube"
                ],
                "summary": "Cubeを削除する (Delete)",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Bearer ??????????",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cube ID",
                        "name": "cube_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/DeleteCubeRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            }
        },
        "/v1/cubes/export": {
            "get": {
                "description": "- USR によってのみ使用できる\n- Cube を .cube ファイルとしてダウンロードする\n- 実行には ExportLimit に残数が必要\n- 成功すると Zip ファイル (application/zip) が返却される",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "v1 Cube"
                ],
                "summary": "Cube をエクスポートする。",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Bearer ??????????",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cube ID",
                        "name": "cube_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cube_uuid.cube",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            }
        },
        "/v1/cubes/genkey": {
            "post": {
                "description": "- USR によってのみ使用できる\n- Exportされた.cubeファイルをアップロードして鍵を発行\n- 発行される鍵には権限と有効期限が含まれる",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "v1 Cube"
                ],
                "summary": "鍵を発行する",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Bearer ??????????",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": ".cube file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CubePermission JSON",
                        "name": "permissions",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "有効期限 (ISO8601)",
                        "name": "expire_at",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/GenKeyCubeRes"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            }
        },
        "/v1/cubes/get/{cube_id}": {
            "get": {
                "description": "- USR によってのみ使用できる\n- Cube の基本情報、統計、系譜情報を取得する",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "v1 Cube"
                ],
                "summary": "Cubeの詳細情報を取得",
                "parameters": [
                    {
                        "type": "string",
                        "example": "Bearer ??????????",
                        "description": "token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Cube ID",
                        "name": "cube_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/GetCubeRes"
                        }
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "AbsorbCubeParam": {
            "type": "object",
            "properties": {
                "chat_model_id": {
                    "type": "integer",
                    "example": 1
                },
                "chunk_overlap": {
                    "type": "integer",
                    "example": 16
                },
                "chunk_size": {
                    "type": "integer",
                    "example": 512
                },
                "content": {
                    "type": "string",
                    "example": "Knowledge base for Go development"
                },
                "cube_id": {
                    "type": "integer",
                    "example": 1
                },
                "half_life_days": {
                    "type": "number",
                    "example": 30
                },
                "is_en": {
                    "type": "boolean",
                    "example": false
                },
                "mdl_k_neighbors": {
                    "type": "integer",
                    "example": 5
                },
                "memory_group": {
                    "type": "string",
                    "example": "legal_expert"
                },
                "min_survival_protection_hours": {
                    "type": "number",
                    "example": 72
                },
                "prune_threshold": {
                    "type": "number",
                    "example": 0.1
                },
                "stream": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "AbsorbCubeRes": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/AbsorbCubeResData"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Err"
                    }
                }
            }
        },
...
    }
}